web/fritzing/requirements.txt
Fabric
Jinja
PyYAML
Pygments
python-dateutil
docutils==0.5
markdown==2.0
textile==2.1.3
yolk==0.4.1
python-memcached
Cheetah

django-dbtemplates==0.5.6
django-templatesadmin==0.5.5
django-reversion==1.1
django-ajax-validation
django-announcements==0.1.0
django-avatar==1.0.1
django-microblogging==0.1.0
django-notification
django-pagination==1.0.5
django-robots==0.6.0
django-messages==0.4.1
django-wikiapp==0.1.2
django-tinymce==1.5
django-imagekit
django-timezones
django-ticker
django-debug-toolbar
django-page-cms
django-bookmarks
South
django-email-confirmation
django-flag==0.1.0
django-friends
django-gravatar
django-mailer
django-oembed
django-threadedcomments
django-locations
django-uni-form
django-swaps
django-extensions
django-markitup
django-db-log
django-licenses
django-filter
vobject

#-e svn+http://svn.pinaxproject.com/pinax/trunk@1612#egg=Pinax
-e git+git://github.com/pinax/pinax.git@0.7beta2#egg=Pinax
-e git+git://github.com/jezdez/django-forum.git#egg=django-forum
-e git+git://github.com/jezdez/dregni.git#egg=dregni
-e git+git://github.com/simonw/django-openid.git#egg=django-openid
-e git+git://github.com/jezdez/django-feedutil.git#egg=django-feedutil
-e svn+http://django-filebrowser.googlecode.com/svn/trunk#egg=django-filebrowser
-e svn+http://django-app-plugins.googlecode.com/svn/trunk#egg=django-app-plugins
-e svn+http://django-tagging.googlecode.com/svn/trunk#egg=django-tagging
-e svn+http://django-voting.googlecode.com/svn/trunk#egg=django-voting
-e svn+http://django-mptt.googlecode.com/svn/trunk#egg=django-mptt
-e svn+http://django-template-utils.googlecode.com/svn/trunk#egg=django-template-utils
-e svn+http://django-emailthis.googlecode.com/svn/trunk#egg=django-emailthis
-e svn+http://django-treemenus.googlecode.com/svn/trunk#egg=django-treemenus
-e hg+http://bitbucket.org/ubernostrum/django-contact-form/#egg=django-contact-form
